requirements.txt
PyYAML~=6.0.1
semver~=3.0.2